i removed and moved a lot of things. Many things are put in archive, confirming before removal. Classifier code is in archive. pending removal.
I only want to keep what is relevant. And i want to organise files into scripts and utility files so that they can more easily be accessed and understood.

scripts have been updated so that it is easier to know what to change or not change.

I would later like to make this a single yaml file that sets all the configurations that can just be imported to each of the files.